Refactor the app initialisation flow
This is the fourth or fifth iteration of this refactor, and definitely
feels the best so far. I've moved all the initialisation stuff and the
URL management to a new `Root` component that chooses what to render.

Now instead of manually doing the `setTimeout` with the
`startListeningToAuth` "action" (which isn't even an action) the `Root`
component takes care of all of that. This means everything now happens
in React instead of some random JS, which feels... better, for the lack
of a more fitting work.

Whatcha think?