Switch to static meta tags
It turns out neither facebook nor Twitter support dynamic meta tags
rendered from JS, so I've hardcoded some meta tags. This is a
non-optimal solution, as it means we can't show the currenty
story/frequency title and description in the preview, but it's good
enough for now.

We need to look into https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md#generating-dynamic-meta-tags-on-the-server in the future.