Properly upgrade users using http cloud function
This commit updates a lot of code. When a user enters their credit card information, it sends a token and plan name to a cloud function. This function will create a customer in Stripe, create a new subscription for that customer with the passed plan, and then return necessary data for us to store in the db (customerId, subscriptionId, tokenId, and customerEmail - in case their payment email is different from their user email)

When this data gets back to the app, we run an update in the db adding the stripe metadata into the users_private object, and adding a small plan field to the user object.

When this write is successful, we fetch the new user, return the snapshot and update the store.

At this point in time nothing happens in the app after the subscription is successful. Next commits will change this to close the upgrade modal, add some sort of visual feedback that the subscription was successful, and apply a (pro) badge to the user across the app