Sort threads by last activity
Adds a new field to threads called `lastActive` which is set whenever a
new message is sent. (to whenever that message was sent)

All old threads will just be sorted by `createdAt` and from the time we
deploy this change when somebody sends a message a thread will be back
at the top of the list! I don't think it's worth the effort to backfill
this.

Closes #1021