Check for existing permissions on channels + communities
On the 'createMemberInX' functions, we now always check for an existing record first in order to decide whether to update it or insert a new record.

This is because we are switching to a non-destructive leave channel and leave community system where instead of deleting those records we now simply mark the user as `isMember: false`

As a result, whenever a user joins a channel or community we now need to check if a previous relationship existed (in which case we just update their membership status), otherwise we create a new membership record.