Speed up user community permission query
Looking at our Apollo Optics data, some of the standout slow parts of our queries are:

- `message.sender`, avg. duration 724ms
- `thread.creator`, avg. duration 557ms

At first glance, it seems like those should be fast, right?
`message.sender` is an id, so just do a `getUser(id)` and done, why
does that take 724ms?! The same thing for `thread.creator`, why does that
take 557ms?!

Looking at the resolvers, much become clear. Due to the complexity in
our setup we have to check permissions before returning anything to the
client.

In the case of `message.sender` this means two additional db
queries:

1. Load `message.thread` to get the `message.thread.communityId`
2. Check the users permission in `message.thread.communityId`

In the case of `thread.creator`, this means one additional query to get
the permission in `thread.community`.

Looking at the first case, getting `message.thread` is (again) only a
`getThread(threadId)`, so a very quick query on the primary key.

**This leads me to believe that our `getUserPermissionInCommunity` query
is part of the problem for our slow loading for signed in users.**

This change fixes the performance issue by introducing a new index on
`usersCommunities`: `userIdAndCommunityId`. By leveraging a `.between`
query in our `getUserPermissionInCommunity` we significantly speed up
the query, especially in large communities, as we avoid a `.filter`
through all `usersCommunities` records in a community.

This requires a migration to be run after merging.