Load messages faster
For ever message loaded we would load the thread that message was in,
again. I.e. if a thread has 100 messages we query for the thread 101x
from the database.

Fixed by using a loader, and I also parallelized the loader work between
fetching the sender and fetching the thread by putting the direct
message thread case first, which should speed up things even more.