Add sCU to chatInput, message and navbar
This fixes typing being slow in long threads. Closes #1462

I went ahead and used react-addons-perf and `Perf.printWasted()` to
figure out what was being slow when sending a message.

The main culprits, unsurprisingly, were the message component (which is
why it's slow when it's a long thread) and the navbar component. (which
should almost never change anyway)

I also added a sCU to the chatinput for good measure since that would
otherwise re-render constantly in the inbox view, unnecessarily.