Fix flashing favicon
Due to a misunderstanding of how `react-helmet` works I had to do an
ugly hack to make sure that our dynamic favicon updates correctly. That
hack resulted in the favicon flashing every time it would update, which
is suboptimal to say the least.

This patch fixes the dynamic favicon once and for all by correctly using
`react-helmet`.

The issue was that `react-helmet` only renders the latest version of a
certain tag if there's duplicates. For example, let's say you have
ComponentA which renders `<link rel="icon" href="componenta.ico" />` and
you have ComponentB which renders `<link rel="icon" href="componentb.ico" />`.

Depending on the order they're rendered in either `href="componenta.ico"` or `href="componentb.ico"` will be actually rendered to the browser DOM.

Due to us using a reusable `Head` component everytime we used that
component it would render a link to the favicon. In hindsight, it makes
total sense that this would break the favicon, as that means whatever is
the last component to render determines which favicon is selected: As
only the navbar actually knows if the unread favicon should be rendered
it was highly unlikely that the unread favicon would show at all. (I'm
kind of surprised that it even worked 50% of the time tbh)

This patch resolves it by only rendering any kind of favicon link from
the navbar, meaning nothing else will override it.

Closes #1865, this will likely conflict with @brianlovin's ongoing work
in #2010 (sorry Brian!) I was just too excited that I finally figured
out why this didn't work as expected so I just had to knock the patch out.