Cleanup pending message state
Rather than hiding the action indicator when a message isn't yet
persisted to the server, we now only hide the "delete" button since you
can't delete a message that hasn't persisted yet.

This means it's much less obvious that a message wasn't sent yet unless
you hover over it and check if the delete button is there, making the
mutation feel much much faster.

Also did a bunch of cleanup of passing that prop around and computing it
again when we weren't even using it anymore in any of the
view-components apart from the actions.