Custom ServiceWorker caching configuration
Based on a recent error I happened to see in my browser console by
chance, I think our ServiceWorker woes come from the fact that the
`offline-plugin` aborts caching anything on install if downloading even
a single asset fails: https://imgur.com/a/q4ut6xx (note how it says
`if (failAll)`, which is true if an asset is in the `main` cache
section—which all assets are by default)

This patch changes that behavior by only failing caching entirely if
either the app shell or the main JS bundles fails to download.
Everything else is downloaded too, but if it fails to download just
isn't cached and instead requested from the server when it's used. (and
then cached if it downloads fine)

Let's hope this fixes our SW woes!